# specs/007-frontend-react-migration/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Dev Debug Platform API
  description: |
    REST API for Dev Debug Platform React SPA.

    **Authentication**: Session-based using Spring Security
    **Base URL**: `/api`
    **WebSocket**: STOMP over WebSocket at `/ws`
  version: 2.0.0
  contact:
    name: CMICT Internal PaaS Team
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://platform.example.com/api
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidate session and log out user
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh session
      description: Extend session expiry time
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Session refreshed
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Servers ====================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
  schemas:
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        isAuthenticated:
          type: boolean
        permissions:
          type: array
          items:
            type: string

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        displayName:
          type: string
        avatar:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
    UserRole:
      type: string
      enum: [SUPER_ADMIN, ADMIN, DEVELOPER]
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, auto]
        language:
          type: string
          enum: [en, zh]
        sidebarCollapsed:
          type: boolean
        defaultDashboard:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            message: Validation failed for field 'name'
            timestamp: '2025-01-04T10:30:00Z'
            path: /api/servers
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Session expired. Please log in again.
            timestamp: '2025-01-04T10:30:00Z'
            path: /api/servers
...
security:
  - sessionAuth: []